requirements.txt
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.32
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.2
pyodbc==5.1.0
